Retrieving specialties from the database
I added a count() function to Specialization Class that counts the specialties stored in the database. These will be added to a drop-down list to be presented to the user.
I added also the feature that saves the patient information into the database after being registered in the Queue Management System.